Ended thread result through server
