+++ LAB3/app/src/main/res/layout/activity_main.xml
@@ -225,7 +225,7 @@
             android:layout_width="match_parent"
             android:layout_height="0dp"
             android:layout_weight=".75"
-            android:onClick="progressBarHandler"
+            android:onClick="callServer"
             android:text="@string/Equal"
             android:visibility="invisible" />
